Restructure ui/src/components folder to have directories for each component
