+++ packages/ui/src/components/Chip/index.ts
@@ -1,0 +1,1 @@
+export * from './Chip';
