+++ packages/ui/src/components/FormHelperText/index.ts
@@ -1,0 +1,1 @@
+export * from './FormHelperText';
