+++ packages/ui/src/components/InputAdornment/index.ts
@@ -1,0 +1,1 @@
+export * from './InputAdornment';
