+++ packages/ui/src/components/InputLabel/InputLabel.stories.tsx
@@ -1,8 +1,8 @@
 import { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
 
-import { FormControl } from './FormControl';
-import { Input } from './Input';
+import { FormControl } from '../FormControl/FormControl';
+import { Input } from '../Input/Input';
 import { InputLabel } from './InputLabel';
 
 const meta: Meta<typeof InputLabel> = {
