+++ packages/ui/src/components/InputLabel/index.ts
@@ -1,0 +1,1 @@
+export * from './InputLabel';
